+++ src/pl/plajer/buildbattle3/handlers/InventoryManager.java
@@ -156,6 +156,7 @@ public class InventoryManager {
         player.setFireTicks(invConfig.getInt("Fire ticks"));
         player.setGameMode(GameMode.valueOf(invConfig.getString("GameMode")));
         player.setAllowFlight(invConfig.getBoolean("Allow flight"));
+        player.setLevel(invConfig.getInt("Exp"));
         List<String> activePotions = invConfig.getStringList("Active potion effects");
         for (String potion : activePotions) {
           String[] splited = potion.split("#");
@@ -172,6 +173,5 @@ public class InventoryManager {
     }
 
     player.updateInventory();
-    player.setLevel(Integer.valueOf(inventory.getTitle()));
   }
 }